B.Tech Final Year Project "Folder_Guard" on .NET MVC
This project is aimed at developing a desktop-based and Protects file
and folder from system. Some features of this application are
password-protect, hide, or restrict access to folders, files, and
programs by using cryptography and steganography. It can also restrict
access to the Control Panel commands, prevent downloading of programs
from the Internet, control user access to the external drives, and more.